pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "xylemx"
version = "0.1.0"
description = "Hackathon project for multimodal deforestation detection"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.9",
  "numpy>=1.26",
  "PyYAML>=6.0",
  "rasterio>=1.4",
  "scikit-learn>=1.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
